Add local dependency base-mc-classes in frontend/base
