+++ frontend/base/main.js
@@ -1,0 +1,6 @@
+const _BaseMosquittoProxyClient = require('./_BaseMosquittoProxyClient');
+
+
+module.exports = {
+    BaseMosquittoProxyClient: _BaseMosquittoProxyClient
+}
