+++ frontend/base/package.json
@@ -1,0 +1,18 @@
+{
+    "name": "base-mc-classes",
+    "description": "",
+    "version": "1.0.0",
+    "main": "main.js",
+    "scripts": {
+      "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "repository": {
+    },
+    "keywords": [],
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+        "axios": "^0.27.2",
+        "uuid":  "^8.3.0"
+    }
+  }
