+++ frontend/package.json
@@ -23,6 +23,7 @@
 		"@material-ui/icons": "^4.11.3",
 		"@material-ui/lab": "^4.0.0-alpha.61",
 		"ajv": "^7.1.1",
+		"base-mc-classes": "file:./base",
 		"brace": "^0.11.1",
 		"express": "^4.18.1",
 		"file-saver": "^2.0.5",
@@ -47,6 +48,7 @@
 		"chart.js": "^2.9.3",
 		"clsx": "latest",
 		"cross-env": "^7.0.3",
+		"form-data": "^4.0.0",
 		"jest": "24",
 		"run-script-os": "^1.1.6"
 	},
